Feat:-       get card list from api and check if error get      from cash and render it.       remove card from api and check if error remove      from cash

+++ app/src/main/java/com/example/payskytask/core/data/model/CardRequestBody.kt
@@ -4,7 +4,7 @@ data class CardRequestBody(
     val cardNumber: String,
     val expireDate: String,
     val cardHolder: String,
-    val balance: Float = 0.00f,
+    val balance: Double = 0.00,
     val logo: String,
     val cvv: String,
 )
